fa2023/abet/pyproject.toml
[tool.poetry]
name = "abet"
version = "0.1.0"
description = ""
authors = ["John Leonard <jdleonard@vcu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
traits = "^6.4.3"
scipi = "^3.0"
fitz = "^0.0.1.dev2"
frontend = "^0.0.3"
tools = "^0.1.9"
loguru = "^0.7.2"
python-pptx = "^0.6.23"
pymupdf = "^1.23.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
